Doing some gitkraken wierdness
I get a glitch where the repo won't open unless i delete an index file. In "Edited the shop", i didn't hold on to the weird changes it makes. I am going to see what happens if I hold on to them. If this fails, revert to "Edited the shop" it worked fine.